separate input and cell
