+++ client/packages/system/src/Item/Components/ItemVariantSelector/ItemVariantInputCell.tsx
@@ -1,14 +1,7 @@
 import React from 'react';
-import {
-  RecordWithId,
-  CellProps,
-  Typography,
-  Box,
-  textStyles,
-  Tooltip,
-} from '@openmsupply-client/common';
-import { useItemVariants } from '../../api';
-import { ItemVariantSelector } from './ItemVariantSelector';
+import { RecordWithId, CellProps } from '@openmsupply-client/common';
+import { ItemVariantFragment } from '../../api';
+import { ItemVariantInput } from './ItemVariantInput';
 
 interface ItemVariantInputCellProps {
   itemId: string;
@@ -20,38 +13,25 @@ export const ItemVariantInputCell = <T extends RecordWithId>({
   column,
   itemId,
   displayInDoses,
-  dense = false,
 }: CellProps<T> & ItemVariantInputCellProps) => {
   const selectedId = column.accessor({
     rowData,
   }) as string | null;
-  const { data, isLoading } = useItemVariants(itemId);
-  const selected = data?.find(variant => variant.id === selectedId);
 
-  const onVariantSelected = (itemVariantId: string | null) => {
-    column.setter({ ...rowData, itemVariantId });
+  const onVariantSelected = (itemVariant: ItemVariantFragment | null) => {
+    column.setter({
+      ...rowData,
+      itemVariantId: itemVariant?.id ?? null,
+      itemVariant,
+    });
   };
 
   return (
-    <Box display="flex" justifyContent="space-between">
-      <Tooltip title={selected?.name}>
-        <Typography
-          sx={{
-            ...(dense ? textStyles.dense : textStyles.default),
-            overflow: 'hidden',
-            textOverflow: 'ellipsis',
-          }}
-        >
-          {selected?.name}
-        </Typography>
-      </Tooltip>
-      <ItemVariantSelector
-        selectedId={selectedId}
-        variants={data}
-        isLoading={isLoading}
-        onVariantSelected={onVariantSelected}
-        displayInDoses={displayInDoses}
-      />
-    </Box>
+    <ItemVariantInput
+      selectedId={selectedId}
+      itemId={itemId}
+      displayInDoses={displayInDoses}
+      onChange={onVariantSelected}
+    />
   );
 };
