+++ client/packages/system/src/Item/Components/ItemVariantSelector/index.ts
@@ -1,1 +1,2 @@
 export { ItemVariantInputCell } from './ItemVariantInputCell';
+export { ItemVariantInput } from './ItemVariantInput';
